Updated DeleteLinkModal Delete Link modal handles deleting a link now, cleaned up LinkCard
